Fix QEMU headless boot: add serial console support
- GRUB gfxterm incompatible with headless mode: switched to
  serial+console dual terminal output
- Kernel had no serial output: added COM1 (0x3F8) serial driver
  that mirrors all _putchar output to serial port
- Fixed inb/outb port width from uint8_t to uint16_t (was truncating
  16-bit port addresses like COM1 0x3F8)
- Mouse ISR could block forever on ps2_controller_read_data() when no
  data available: fixed to check buffer status first
- PS/2 busy-wait loops had no timeout: added timeout guards to all
  PS/2 controller I/O functions
- QEMU run script defaults to -nographic, added --gui flag
- GRUB timeout reduced to 1s for faster headless boot
- Boot progress messages output to serial during init

+++ include/kernel/kernel.h
@@ -53,8 +53,8 @@ extern "C" {
 	extern uint32_t isr_division_by_zero;	// division by zero interrupt handler
 	extern uint32_t isr_mouse;		// mouse interrupt handler - wraps isr_mouse_handler
 	extern uint32_t isr;	
-	extern uint8_t inb(uint8_t reg);		// get byte from memory register reg
-	extern void outb(uint8_t reg, uint8_t byte);	// write byte to memory register reg
+	extern uint8_t inb(uint16_t reg);		// get byte from I/O port
+	extern void outb(uint16_t reg, uint8_t byte);	// write byte to I/O port
 	
 #ifdef __cplusplus
 }
